requirements.txt
mss==9.0.1
opencv-python==4.10.0.84
numpy==2.1.2
pyautogui==0.9.54
pytesseract==0.3.13
pillow==10.4.0
pyyaml==6.0.2
ttkbootstrap==1.10.1
pynput==1.7.7
